add mac builds with libzvbi as an additional matrix configuration
enable it for macos-13/intel and macos-15/arm, but exclude it from the
macos-14 job.
Besides enabling libzvbi also add dvd support, and enable the rest of
the features to closely match the binary version from homebrew core,
even when they are not really useful in the context of mpv (like e.g.
libtesseract)